Fixed warnings

+++ stingraykit/time/NtpTimestamp.cpp
@@ -15,7 +15,7 @@ namespace stingray
 		u64 seconds = integerPart - DifferenceBetweenUnixNtpTime;
 
 		u32 fractionPart = _timestamp - integerPart;
-		u64 milliSeconds = fractionPart / (1UL << 32);
+		u64 milliSeconds = fractionPart / ((u64)1 << 32);
 
 		return Time(seconds * 1000 + milliSeconds);
 	};
@@ -24,7 +24,7 @@ namespace stingray
 	NtpTimestamp NtpTimestamp::FromTime(const Time& time)
 	{
 		u32 integerPart = time.GetMilliseconds() / 1000 +  DifferenceBetweenUnixNtpTime;
-		u32 fractionPart = time.GetMilliseconds() / 1000 * (1UL << 32);
+		u32 fractionPart = time.GetMilliseconds() / 1000 * ((u64)1 << 32);
 
 		u64 timestamp = ((u64)integerPart << 32) | fractionPart;
